Separate Nisa parser from core parser-service
Core parser service now consumes via the new modules
